tests/wpt/mozilla/tests/mozilla/PutForwards.html

<!doctype html>
<meta charset="utf-8">
<title></title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  assert_equals(window.location.href, window.location.toString());
  assert_equals(document.location.href, document.location.toString());
}, "Forwarded attributes are handled correctly");
</script>
